+++ src/Soluble/DbWrapper/Exception/UnsupportedFeatureException.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Soluble\DbWrapper\Exception;
+
+class UnsupportedFeatureException extends \RuntimeException implements ExceptionInterface
+{
+}