+++ src/Soluble/DbWrapper/Result/ResultInterface.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Soluble\DbWrapper\Result;
+
+interface ResultInterface
+{
+};