+++ src/Soluble/DbWrapper/Result/Resultset.php
@@ -4,7 +4,7 @@ namespace Soluble\DbWrapper\Result;
 
 use ArrayObject;
 
-class Resultset extends ArrayObject
+class Resultset extends ArrayObject implements ResultInterface
 {