+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
-### 1.0.0 (2016-01-02) First stable release
+## 1.1.0 (2016-01-06) Universal drivers
+
+- Added support for zend-db, laravel and doctrine
+
+## 1.0.0 (2016-01-02) First stable release